+++ qb-engineer.core/Entities/ShipmentLine.cs
@@ -5,7 +5,9 @@ public class ShipmentLine : BaseEntity
     public int ShipmentId { get; set; }
     public int? SalesOrderLineId { get; set; }
     public int? PartId { get; set; }
-    public int Quantity { get; set; }
+    // Phase 3 / WU-23 (F8-broad): UoM-aware fractional quantities. Promoted to
+    // decimal(18,4) so a fractional SO line can ship without truncation.
+    public decimal Quantity { get; set; }
     public string? Notes { get; set; }
     public string? Description { get; set; }
     public decimal? Weight { get; set; }
